added drop table precondition to handle transition from order to orders table on migrate:refresh.

+++ database/migrations/2016_10_24_014307_create_order_table.php
@@ -32,6 +32,13 @@ class CreateOrderTable extends Migration
      */
     public function down()
     {
-        Schema::drop('order');
+        /* temporary till everyone has refreshed database tables */
+        if(Schema::hasTable('order')){
+            Schema::drop('order');
+        }
+        
+        if(Schema::hasTable('orders')){
+            Schema::drop('orders');
+        }
     }
 }
